relocate queries to app and navigation provider
