+++ src/App.js
@@ -13,11 +13,45 @@ import { NavigationProvider } from './Context/NavigationProvider.js';
 import { SchoolProvider } from './Context/SchoolProvider.js';
 import WelcomePage from './Pages/WelcomePage.js';
 import RegistrationPage from './Pages/RegistrationPage.js';
-import RestService from './Services/RestService.js';
+import axios from 'axios';
+
+//Function that allows us to accept credentials
+const instance = axios.create({
+  baseURL: 'http://localhost:4000', // Set your backend URL
+  withCredentials: true, // Enable sending cookies with cross-origin requests
+});
 
 function App() {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [isSuperAdmin, setIsSuperAdmin] = useState(false);
+
+  const validateToken = async (token) => {
+    try {
+      if (token) {
+        const response = await instance.get(`${process.env.REACT_APP_API_URL_AUTH}/verify/?token=${token}`)
+        if (response) {
+          console.log(response.data)
+        }
+        return response.data
+      }
+    } catch (error) {
+      console.error('Error validating token:', error);
+      throw new Error("Token validation failed. Please try again later.");
+    }
+  };
+
+  const getUserById = async (user_id) => {
+    try {
+      const response = await instance.get(`${process.env.REACT_APP_API_URL_USER}/${user_id}`)
+      if (response) {
+        console.log(response.data);
+      }
+      return response.data;
+    } catch (error) {
+      console.error('Error fetching user:', error);
+      throw new Error("Get user failed. Please try again later.");
+    }
+  };
 
   useEffect(() => {
     const fetchData = async () => {
@@ -30,11 +64,11 @@ function App() {
           const token = jwtCookie.split('=')[1];
           console.log('JWT Token:', token);
 
-          // Call RestService to validate the token
-          const data = await RestService.validateToken(token);
+          // Call to validate the token
+          const data = await validateToken(token);
 
           if (data) { // access data.id
-            const user = await RestService.getUserById(data.id);
+            const user = await getUserById(data.id);
             user.position === "Super administrator" && setIsSuperAdmin(true); //is admin
             setIsLoggedIn(true)
           } else {
